feat: 클래스 등록 전체 form 버튼 활/비활성화 로직 (#206)
