+++ src/pages/instructor/classRegister/index.tsx
@@ -7,7 +7,6 @@ import Header from '@/components/Header';
 import useBottomSheet from '@/hooks/useBottomSheet';
 import useImageUploader from '@/hooks/useImageUploader';
 import { useGetLocationList, usePostClassRegisterInfo } from '@/apis/instructor/classRegister/queries';
-import { formatToISOString } from '@/utils/timeUtils';
 import { genreEngMapping, levelEngMapping } from '@/constants';
 import ClassAmount from './ClassAmount';
 import ClassDescription from './ClassDescription';
@@ -26,13 +25,6 @@ const ClassRegister = () => {
   const { isBottomSheetOpen, openBottomSheet, closeBottomSheet } = useBottomSheet();
   const [selectedLocation, setSelectedLocation] = useState<LocationTypes | null>(null);
 
-  const [times, setTimes] = useState<{ startTime: string; endTime: string; date: string; duration: number }[]>([]);
-  const [startDate, setStartDate] = useState('');
-  const [hour, setHour] = useState(12);
-  const [minute, setMinute] = useState(0);
-  const [ampm, setAmpm] = useState('AM');
-  const [selectedTime, setSelectedTime] = useState<number | null>(null);
-
   const { mutate: classRegisterMutate } = usePostClassRegisterInfo();
   const {
     explainTextAreaRef,
@@ -49,10 +41,22 @@ const ClassRegister = () => {
     submitDefaultPlace,
     detailPlace,
     amount,
+    selectedTime,
+    times,
+    hour,
+    minute,
+    ampm,
+    startDate,
+    setStartDate,
     handleClassNameChange,
     handlePersonnelChange,
     handleAmountChange,
     setImageUrls,
+    setHour,
+    setMinute,
+    setAmpm,
+    setSelectedTime,
+    handleAddTime,
     toggleCategory,
     handleLevelSelect,
     handleExplainTextArea,
@@ -61,6 +65,8 @@ const ClassRegister = () => {
     handleDefaultPlace,
     handleSubmitDefaultPlace,
     handleDetailPlace,
+    handleRemoveTime,
+    isFormValid,
   } = useClassRegisterForm();
 
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
@@ -90,25 +96,6 @@ const ClassRegister = () => {
 
       classRegisterMutate(updatedInfo);
     }
-  };
-
-  const handleAddTime = () => {
-    if (startDate && selectedTime !== null) {
-      const { startTime, endTime } = formatToISOString(startDate, hour, minute, ampm, selectedTime);
-
-      const newTime = {
-        startTime,
-        endTime,
-        date: startDate,
-        duration: selectedTime,
-      };
-      setTimes([newTime, ...times]);
-    }
-  };
-
-  const handleRemoveTime = (index: number) => {
-    const updatedTimes = times.filter((_, idx) => idx !== index);
-    setTimes(updatedTimes);
   };
 
   const handleImageUploadSuccess = (url: string) => {
@@ -168,7 +155,9 @@ const ClassRegister = () => {
         </div>
 
         <div className={buttonContainerStyle}>
-          <BoxButton type="submit">완료</BoxButton>
+          <BoxButton type="submit" disabled={!isFormValid()}>
+            완료
+          </BoxButton>
         </div>
       </form>
       {isBottomSheetOpen && (
